Added redshift threshold for hmf in utils.py + test cov for hm and example notebook
