--- soliket/halo_model/halo_model_fe/HaloModel_fe.yaml
@@ -1,32 +1,0 @@
-read_matterPS: False #if True, reads pre-computed linear matter PS
-                        #if False, computes the linear matter PS using CAMB
-gal_mod: False
-matter_matter: 'mm'
-galaxy_galaxy: #'gg'
-matter_galaxy: "gm"
-redshift: './tabulated/redshift.txt'
-
-z : [0,1,2]
-
-kmax: 10.0
-
-Mmin: 11
-Mmax: 15
-nm: 200 
-
-Dc: 200
-
-T_CMB: 2.725
-tau: 0.0544
-ns: 0.9649
-As: 1.97448e-9 #2.101e-9
-pivot_scalar: 0.05
-matter_PS:
-sigma_EP: 0.1
-sigma_LP: 0.1
-scale_EP: 10.0
-scale_LP: 10.0
-alpha_EP: 1.0
-alpha_LP: 1.0
-LogMmin_EP: 12.0
-LogMmin_LP: 10.8
