--- soliket/halo_model/halo_model_fe/__init__.py
@@ -1,5 +1,0 @@
-from .halo_model_fe import HaloModel_fe
-
-__all__ = {
-    "HaloModel_fe"
-}
