+++ soliket/halo_model_fe/__init__.py
@@ -1,0 +1,6 @@
+from .halo_model_fe import HaloModel, HaloModel_fe
+
+__all__ = {
+    "HaloModel",
+    "HaloModel_fe"
+}
