+++ soliket/halo_model_fe/tabulated/redshift.txt
@@ -1,0 +1,53 @@
+0.001000000000000000
+4.954242706298828125e-02
+9.395635128021240234e-02
+1.402497291564941406e-01
+1.885021924972534180e-01
+2.387965917587280273e-01
+2.912192344665527344e-01
+3.458603620529174805e-01
+4.028136730194091797e-01
+4.621771574020385742e-01
+5.240527391433715820e-01
+5.885467529296875000e-01
+6.557699441909790039e-01
+7.258379459381103516e-01
+7.988709211349487305e-01
+8.749945163726806641e-01
+9.543395042419433594e-01
+1.037042140960693359e+00
+1.123244524002075195e+00
+1.213094711303710938e+00
+1.306747198104858398e+00
+1.404362916946411133e+00
+1.506109237670898438e+00
+1.612161397933959961e+00
+1.722701311111450195e+00
+1.837918996810913086e+00
+1.958012580871582031e+00
+2.000000000000000000
+2.083187818527221680e+00
+2.213660717010498047e+00
+2.349654197692871094e+00
+2.491403102874755859e+00
+2.639150381088256836e+00
+2.793149709701538086e+00
+2.953666210174560547e+00
+3.120975494384765625e+00
+3.295364379882812500e+00
+3.477133274078369141e+00
+3.666594028472900391e+00
+3.864071846008300781e+00
+4.069906711578369141e+00
+4.284452438354492188e+00
+4.508077144622802734e+00
+4.741164684295654297e+00
+4.984116077423095703e+00
+5.000000000000000000
+5.237348556518554688e+00
+5.501297473907470703e+00
+5.776415348052978516e+00
+6.063176155090332031e+00
+6.362070560455322266e+00
+6.673614501953125000e+00
+
